More robust integration test (#1430)
* [e2e-tests] Add waitForNthState helper

* Increase size of stub file by 10x

This gives the test more "space" to eg. wait for 50 turns.

* [e2e-test] Wait for 10 turns before cancelling

* [e2e] Only close browsers in headless mode

This lets us inspect the state of the browsers on completion/failure

* Use different provider signing keys (to get distinct destinations)

* Fix puppeteer test unsubscribing all ChannelUpdated subscribers

* Fix undefined issue

There's an issue where this.paymentChannelClient.channelCache[channelId]
is sometimes undefined. I believe that the channelCache is updated in
a different event loop, which isn't guaranteed to run first.

* Avoid the spread operator in insertIntoChannelCache

* Always assign when calling updateChannelCache

* Don't log from web3torrent in puppeteer test

This causes some random DOM exception event

* Rename channelId

Currently, releaseFunds is called with the target channel id. While this
change does not prevent this mistake, it serves as documentation.

* Don't call releaseFunds in concludeChannel

Funds should be released when the ledger channel's outcome is updated --
until that point, the funds are still unavailable.

Thus, releaseFunds should be called by virtualDefundingAsLeaf